+++ automotive/src/main/java/au/com/shiftyjelly/pocketcasts/AutomotiveLinkFragment.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.fragment.app.Fragment
 import au.com.shiftyjelly.pocketcasts.compose.AutomotiveTheme
-import au.com.shiftyjelly.pocketcasts.compose.rememberQrPainter
+import au.com.shiftyjelly.pocketcasts.qr.rememberQrPainter
 
 /**
  * Fallback for when the Automotive platform doesn't have a web browser.
