+++ settings.gradle.kts
@@ -84,6 +84,7 @@ include(":modules:services:model")
 include(":modules:services:payment")
 include(":modules:services:preferences")
 include(":modules:services:protobuf")
+include(":modules:services:qr")
 include(":modules:services:repositories")
 include(":modules:services:sharing")
 include(":modules:services:servers")
